# Swagger.yaml
swagger: '2.0'
info:
  description: 'Clone da Api do projeto de Vendas do Professor Dougllas Sousa (http://github.com/cursodsousa)'
  version: 1.0.1
  title: Vendas API
  contact:
    name: Denilson Marcio
    url: 'https://github.com/DenilsonMarcio'
    email: denilson.santos7@fatec.sp.gov.br
host: 'localhost:8080'
basePath: /
...
paths:
...
  /api/v1/usuarios:
    post:
      tags:
        - usuario-controller
      summary: Criar um usuário
      operationId: salvarUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: usuario
          description: usuario
          required: true
          schema:
            $ref: '#/definitions/Usuario'
      responses:
        '201':
          description: Usuário salvo com sucesso
          schema:
            $ref: '#/definitions/Usuario'
        '400':
          description: Erro de validação
      security:
        - JWT:
            - global
      deprecated: false
  /api/v1/usuarios/auth:
    post:
      tags:
        - usuario-controller
      summary: Autenticar um Usuario
      operationId: autenticarUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: credenciais
          description: credenciais
          required: true
          schema:
            $ref: '#/definitions/CredenciaisDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/TokenDTO'
        '400':
          description: Erro de validação
      security:
        - JWT:
            - global
      deprecated: false
  '/api/v1/usuarios/reset/{id}/{login}':
    patch:
      tags:
        - usuario-controller
      summary: Resetar senha do Usuario
      operationId: resetaSenhaUsingPATCH
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: dto
          description: dto
          required: true
          schema:
            $ref: '#/definitions/UsuarioResetaSenhaDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
        - name: login
          in: path
          description: login
          required: true
          type: string
      responses:
        '200':
          description: OK
        '401':
          description: Não autorizado
        '403':
          description: Token não autenticado
      security:
        - JWT:
            - global
      deprecated: false
  /api/v1/usuarios/users:
    get:
      tags:
        - usuario-controller
      summary: Listar todos os Usuarios
      operationId: findUsingGET_2
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Usuario'
        '401':
          description: Não autorizado
        '403':
          description: Token não autenticado
      security:
        - JWT:
            - global
      deprecated: false
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  CredenciaisDTO:
    type: object
    properties:
      login:
        type: string
      senha:
        type: string
    title: CredenciaisDTO
...
  TokenDTO:
    type: object
    properties:
      login:
        type: string
      token:
        type: string
    title: TokenDTO
  Usuario:
    type: object
    properties:
      admin:
        type: boolean
      id:
        type: integer
        format: int32
      login:
        type: string
      senha:
        type: string
    title: Usuario
  UsuarioResetaSenhaDTO:
    type: object
    properties:
      novaSenha:
        type: string
    title: UsuarioResetaSenhaDTO